requirements.txt
Django==5.0.6
psycopg[binary]==3.1.18
dj-database-url==2.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
djangorestframework==3.15.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
whitenoise==6.7.0



